+++ packages/perseus-editor/src/widgets/__docs__/image-editor.stories.tsx
@@ -84,6 +84,8 @@ export const Populated: Story = {
     args: {
         backgroundImage: {
             url: "https://cdn.kastatic.org/ka-content-images/61831c1329dbc32036d7dd0d03e06e7e2c622718.jpg",
+            width: 400,
+            height: 225,
         },
         alt: "The moon showing behind the Earth in space.",
         caption: "Captured via XYZ Telescope",