+++ packages/perseus-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @khanacademy/perseus-core
 
+## 18.9.0
+
+### Minor Changes
+
+-   [#2799](https://github.com/Khan/perseus/pull/2799) [`8b18b38b31`](https://github.com/Khan/perseus/commit/8b18b38b31866582cb101d29ea3dbfd2dbe756bf) Thanks [@anakaren-rojas](https://github.com/anakaren-rojas)! - Changes the output of radio score from boolean array to string array that represent the user's selected choice ids
+
 ## 18.8.0
 
 ### Minor Changes